+++ CHANGELOG.md
@@ -1,5 +1,13 @@
-### NEXT_VERSION_TYPE=MAJOR|MINOR|PATCH
+### NEXT_VERSION_TYPE=MINOR
 ### NEXT_VERSION_DESCRIPTION_BEGIN
+
+* Add `BYTES_PER_SECOND` Unit for YAxis
+* Add the ability to group the values of query variable into selectable tags
+```kotlin
+    val variable by builder.query(datasource = Zabbix, query = "App version") {
+        tags = VariableTags("*.*", "*.\$tag")
+    }
+```
 ### NEXT_VERSION_DESCRIPTION_END
 ## [2.10.0]() (22-11-2019)